+++ src/aiqclib/common/base/config_base.py
@@ -21,14 +21,7 @@ from aiqclib.common.config.yaml_schema import (
     get_classification_config_schema,
     get_nrtqc_config_schema,
 )
-from aiqclib.common.config.yaml_templates import (
-    get_config_data_set_template,
-    get_config_data_set_full_template,
-    get_config_train_set_template,
-    get_config_classify_set_template,
-    get_config_classify_set_full_template,
-    get_config_nrtqc_template,
-)
+from aiqclib.common.config.yaml_templates import get_template_text
 from aiqclib.common.utils.config import get_config_item
 from aiqclib.common.utils.config import read_config
 from aiqclib.common.utils.file import expand_path
@@ -102,18 +95,8 @@ class ConfigBase(ABC):
         if section_name not in yaml_schemas:
             raise ValueError(f"Section name {section_name} is not supported.")
 
-        yaml_templates = {
-            "template:data_sets": get_config_data_set_template,
-            "template:data_sets_full": get_config_data_set_full_template,
-            "template:training_sets": get_config_train_set_template,
-            "template:classification_sets": get_config_classify_set_template,
-            "template:classification_sets_full": get_config_classify_set_full_template,
-            "template:nrt_qc_sets": get_config_nrtqc_template,
-        }
         if str(config_file).startswith("template:"):
-            if str(config_file) not in yaml_templates:
-                raise ValueError(f"Template name {config_file} is not supported.")
-            full_config = yaml.safe_load(yaml_templates.get(str(config_file))())
+            full_config = yaml.safe_load(get_template_text(str(config_file)))
         else:
             full_config = read_config(config_file)